pyproject.toml
[project]
name = "streamlit-study"
version = "0.1.0"
description = "Streamlit入門テンプレートアプリ"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "pillow>=10.0.0",
]
requires-python = ">=3.8.1"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
] 
